requirements.txt
# requirements.txt
tkcalendar
reportlab
win10toast
Pillow